get rid of ReadableBufExtras functions
i don't think we need to be SO flexible that we support reading an
arbitrary number of bits to an arbitrary type.  for now, just support
reading <= 8 bits into a u8.  in the future can add logic to support
crossing byte boundaries in those helpers and reading into u16, etc.